Progressive LLM elimination: offline-first extraction, embeddings, and briefing
- Add extraction provider factory (auto → anthropic → ollama → narrow)
  supporting zero-API-key operation via deterministic narrow extractors
- Add NarrowExtractorAdapter wrapping evidence pipeline behind EntityExtractor interface
- Add OllamaExtractor for local LLM extraction via Ollama API
- Replace LLM briefing synthesis with deterministic template renderer
- Make fastembed a default dependency for offline vector search
- Change embedding provider default from "voyage" to "auto" (voyage → local → noop)
- Wire factory into all 3 instantiation points (main, mcp, cli)
- Remove ANTHROPIC_API_KEY from SKILL.md required env
- Update .env.example, CLAUDE.md, config docs

Evidence pipeline, consolidation, installer, dashboard, website, and
benchmark improvements from recent development cycle included.

2464 tests passing, ruff clean.

+++ server/engram/consolidation/maturity_features.py
@@ -50,7 +50,7 @@ def compute_access_regularity(access_intervals: list[float]) -> float:
         sum((interval - mean_interval) ** 2 for interval in access_intervals)
         / len(access_intervals)
     ) ** 0.5
-    return max(0.0, 1.0 - (std_dev / mean_interval))
+    return float(max(0.0, 1.0 - (std_dev / mean_interval)))
 
 
 def get_cached_maturity_features(
